IsStringUnique with HashTable
a hash table is way faster here, becuase instead of looping throught the array once per character in the array, it only does so once, and stops when/if there is a duplicate in the hash table.

worth noting that this is not even a full hash table, because there is no need for linked lists. If there is a collision, we know there is a duplicate.

this commit also reorganized files