SortStack maintains lowest value on top
SortStack attempts to place the lowest value on the top of the stack. If a larger item is about to be placed on top, we remove the top node and hold it in another temporary stack. We repeat this process until we find the proper place for the new item. Then, we put everything from the temporary stack back on top of the new item.

User Interface warns to user that this app is all in the tests. Hopefully this will minimize confusion in people who pull repo and run.

Reorganized test files. In prior commits, there was one large test file per category of classes; I am shifting to one test class per class.